requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
decorator==5.1.1
Deprecated==1.2.14
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
fabric==3.2.2
Faker==26.3.0
inflection==0.5.1
invoke==2.2.0
Markdown==3.6
packaging==24.2
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
validate-docbr==1.10.0
wrapt==1.16.0
